api/requirements.txt
flask
matplotlib
flask-cors
python-dotenv
gunicorn